ScaleInput tests
